requirements.txt
Flask>=3.0.0
Flask-SocketIO>=5.3.0
geopy>=2.4.0
requests>=2.31.0
eventlet>=0.36.0
gunicorn>=21.2.0
beautifulsoup4>=4.12.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.2.0
python-dotenv>=1.0.1
ijson>=3.3.0
